Add basic construction methods, updatePosition()
The PLObject class represents a object with a size, position, velocity,
and color. Of these, size is mandatory and the rest are optional, with
default values. Unit tests are provided for the build methods.

The method updatePosition() is implemented. It applies the given velocity
to the object's position, based on a given time delta. Unit tests are
provided for basic usage. Corner cases and extreme values are not covered.

Signed-off-by: Patrick Luhmann <patrick.luhmann@gmail.com>