Add bitmap to PLObject, along with more tests
I decided to go with a real bitmap instead of just using a color.
Any user that wants a plain rectangle can just use a boring bitmap.

More unit tests were added. I ran into the "Android API not mocked" issue
and decided that the best solution was to move the affected tests into
the instrumented section. This requires an emulated device, which is annoying,
but it allows the tests to run as desired. I think this is the only way
to handle classes that are used internally by the class under test.

Signed-off-by: Patrick Luhmann <patrick.luhmann@gmail.com>