Implement collision detection method
Method willCollide() tells the caller if and when two objects
will collide. This can be used to handle the collision before it
actually happens, giving the caller complete control.

Unit tests need to be instrumented because of the reliance on the
Rect class, which is empty (in java.jar) and apparently cannot
be mocked in the way that I want.

Signed-off-by: Patrick Luhmann <patrick.luhmann@gmail.com>